requirements.txt
langchain>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-deepseek>=1.0.0
langchain-huggingface>=1.0.0
langchain-tavily>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=1.0.0
faiss-cpu>=1.9.0
gradio>=5.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
numexpr>=2.8.0
tiktoken>=0.7.0
sentence-transformers>=3.0.0
